backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
httpx==0.26.0
authlib==1.6.6
itsdangerous==2.1.2
python-dotenv==1.0.0
celery==5.3.6
redis==5.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
email-validator
PyJWT[crypto]==2.9.0
cryptography>=41.0.0
apscheduler==3.10.4
